Move @Plugin classes to ui.plugin, ZarrUtils to imglib2
